JAVA-10081 :Create new module hibernate-mapping from spring-hibernate-5 module
