+++ persistence-modules/hibernate-many-to-many/README.md
@@ -1,0 +1,7 @@
+## Hibernate 5 with Spring
+
+This module contains articles about Hibernate 5 with Spring.
+
+### Relevant articles
+
+- [Hibernate Many to Many Annotation Tutorial](https://www.baeldung.com/hibernate-many-to-many)
