+++ persistence-modules/hibernate-many-to-many/src/main/resources/persistence-h2.properties
@@ -1,0 +1,21 @@
+# jdbc.X
+jdbc.driverClassName=org.h2.Driver
+jdbc.url=jdbc:h2:mem:db;DB_CLOSE_DELAY=-1
+jdbc.eventGeneratedId=sa
+jdbc.user=sa
+jdbc.pass=
+
+# hibernate.X
+hibernate.dialect=org.hibernate.dialect.H2Dialect
+hibernate.show_sql=false
+hibernate.hbm2ddl.auto=create-drop
+hibernate.cache.use_second_level_cache=true
+hibernate.cache.use_query_cache=true
+hibernate.cache.region.factory_class=org.hibernate.cache.ehcache.EhCacheRegionFactory
+
+# hibernate.search.X
+hibernate.search.default.directory_provider = filesystem
+hibernate.search.default.indexBase = /data/index/default
+
+# envers.X
+envers.audit_table_suffix=_audit_log
