+++ persistence-modules/hibernate-many-to-many/src/test/resources/manytomany.cfg.xml
