--- persistence-modules/spring-hibernate-5/src/main/java/com/baeldung/persistence/manytomany/dao/IProjectDao.java
@@ -1,8 +1,0 @@
-package com.baeldung.persistence.manytomany.dao;
-
-import com.baeldung.hibernate.manytomany.model.Project;
-import com.baeldung.persistence.dao.common.IOperations;
-
-public interface IProjectDao extends IOperations<Project>{
-
-}
